youtube-summarizer/backend/requirements.txt
Flask
Flask-Cors
youtube-transcript-api
# Add other libraries like 'openai' or 'google-generativeai' here later
python-dotenv
google-generativeai
waitress
gunicorn